+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.35.5
 
+* `Core`:
+    * Several new extensions `ByteReadChannel#asStorageFile` and `ByteReadChannelAllocator#asStorageFile`
 * `API`:
     * New extensions `TelegramBot#downloadFile` for writing of incoming bytes to the file
     * New extensions `TelegramBot#downloadFileStream` and `TelegramBot#downloadFileStreamAllocator` for getting of input